Another huge update
Added external LSM library and modified it to work with standardized functions
required by my testbench. Also fixed memory leaks, errors, and other silliness
pre-existing in the library. It passes simple tests without leaks now so
hopefully everything is ok. Performance is around expected range given current
tests. Kept the original source of the LSM tree as a submodule and will cite it
in the report.

Optimized vector: now stores capacity in first index and actual index is now
just computed by index + 1. This resulted in the vector being the fastest for
very basic tests. I imagine this outlook will change when the range of values
becomes larger and memory utilization of the vector likely will be the worst.

Fixed some simple things that were annoying me such as stuck to using uintptr_t
(mostly because it is possible to convert to it easily from (void *))

Next steps:
- actual benchmark (what I have now is simply a testbench)
- make LSM tree persistent (OH BOY, this is gonna be fun)

+++ src/Common.c
@@ -40,7 +40,7 @@ void pmemInit() {
 // sets *POBJ to the direct object pointer
 // example: pmemMalloc(&node, sizeof(SLLNode), SLLNode_TYPE)
 // where node is SLLNode *
-void pmemMalloc(void **POBJ, size_t SIZE, uint32_t TYPE_NUM) { \
+void pmemMalloc(void **POBJ, size_t SIZE, uintptr_t TYPE_NUM) { \
   PMEMoid OID;
   int ret = pmemobj_zalloc(pop, &OID, SIZE, TYPE_NUM);
   if (ret == -1) {
@@ -67,7 +67,7 @@ void pmemDestroy() {
   pmemobj_close(pop);
 }
 
-void pmemPersist(void *OBJ, uint32_t LEN) {
+void pmemPersist(void *OBJ, uintptr_t LEN) {
   pmemobj_persist(pop, OBJ, LEN); \
 }
 
@@ -87,7 +87,7 @@ uint8_t file_exists(const char *path) {
 
 // since Generic is using unsigned integers, hash function just returns their
 // value
-uint32_t GenericHash(Generic data) {
+uintptr_t GenericHash(Generic data) {
   return (uintptr_t) data;
 }
